Add files via upload
FORECASTING INFLATION IN INDIA
Abstract: 
India is an emerging and developing economy. Inflation Index is one of the most important parameters of the Indian Economy. Hence, it becomes important to predict the inflation index as it can help the central bank to decide the best of the monetary policy to achieve price stability. The main purpose of this study is to forecast inflation rates in the case of the Indian economy with the help of SARIMAX model and the NKPC economic model. Our objective is to use both the models to predict or forecast inflation index of India for the period: 2021-2060. Another objective is to compare which of these two models are better to almost predict the inflation index accurately.

Keywords: Forecasting, Emerging economies, Inflation, SARIMAX model, Keynesian Phillips curve.